Rename clutter service
